requirements.txt
Django==1.5
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.1
django-tastypie==0.9.12
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
six==1.2.0
wsgiref==0.1.2
